requirements.txt
httpx
funcy
prefect
jsonlines
google-cloud-storage
pydantic
pendulum